Bump glslang and deps
This may need more testing

I'm not sure if casting `glslang_shader_options_t` and
`glslang_messages_t` like this is the best way to do it

+++ glslang/src/program.rs
@@ -56,7 +56,7 @@ impl<'a> Program<'a> {
             | glslang_sys::glslang_messages_t::VULKAN_RULES
             | glslang_sys::glslang_messages_t::SPV_RULES;
 
-        if unsafe { sys::glslang_program_link(self.handle.as_ptr(), messages.0) } == 0 {
+        if unsafe { sys::glslang_program_link(self.handle.as_ptr(), messages.0 as i32) } == 0 {
             return Err(GlslangError::LinkError(self.get_log()));
         }
         Ok(())
@@ -79,7 +79,7 @@ impl<'a> Program<'a> {
             | glslang_sys::glslang_messages_t::VULKAN_RULES
             | glslang_sys::glslang_messages_t::SPV_RULES;
 
-        if unsafe { sys::glslang_program_link(self.handle.as_ptr(), messages.0) } == 0 {
+        if unsafe { sys::glslang_program_link(self.handle.as_ptr(), messages.0 as i32) } == 0 {
             return Err(GlslangError::LinkError(self.get_log()));
         }
 
@@ -116,7 +116,7 @@ impl<'a> Program<'a> {
             | glslang_sys::glslang_messages_t::VULKAN_RULES
             | glslang_sys::glslang_messages_t::SPV_RULES;
 
-        if unsafe { sys::glslang_program_link(self.handle.as_ptr(), messages.0) } == 0 {
+        if unsafe { sys::glslang_program_link(self.handle.as_ptr(), messages.0 as i32) } == 0 {
             return Err(GlslangError::LinkError(self.get_log()));
         }
